+++ src/main/gen/io/github/facilityapi/intellij/psi/FsdDecoratedField.java
@@ -5,7 +5,7 @@ import java.util.List;
 import org.jetbrains.annotations.*;
 import com.intellij.psi.PsiElement;
 
-public interface FsdDecoratedField extends PsiElement {
+public interface FsdDecoratedField extends FsdDecoratedElement {
 
   @NotNull
   List<FsdAttributeList> getAttributeListList();