+++ src/main/gen/io/github/facilityapi/intellij/psi/FsdDecoratedServiceItem.java
@@ -5,7 +5,7 @@ import java.util.List;
 import org.jetbrains.annotations.*;
 import com.intellij.psi.PsiElement;
 
-public interface FsdDecoratedServiceItem extends PsiElement {
+public interface FsdDecoratedServiceItem extends FsdDecoratedElement {
 
   @NotNull
   List<FsdAttributeList> getAttributeListList();