+++ src/main/kotlin/io/github/facilityapi/intellij/intention/FieldValidateIntention.kt
@@ -1,9 +1,7 @@
 package io.github.facilityapi.intellij.intention
 
 import com.intellij.codeInsight.intention.PsiElementBaseIntentionAction
-import com.intellij.codeInsight.template.Template
 import com.intellij.codeInsight.template.TemplateManager
-import com.intellij.codeInsight.template.impl.TemplateSettings
 import com.intellij.openapi.editor.Editor
 import com.intellij.openapi.project.Project
 import com.intellij.psi.PsiDocumentManager
@@ -12,13 +10,13 @@ import com.intellij.psi.PsiWhiteSpace
 import com.intellij.psi.util.PsiTreeUtil
 import com.intellij.psi.util.siblings
 import io.github.facilityapi.intellij.FsdBundle
-import io.github.facilityapi.intellij.FsdLanguage
+import io.github.facilityapi.intellij.getValidateTemplateForField
 import io.github.facilityapi.intellij.psi.FsdAttributeList
 import io.github.facilityapi.intellij.psi.FsdDecoratedField
 import io.github.facilityapi.intellij.psi.FsdEnumSpec
-import io.github.facilityapi.intellij.psi.FsdField
 import io.github.facilityapi.intellij.reference.addAttribute
 import io.github.facilityapi.intellij.reference.createFromText
+import io.github.facilityapi.intellij.supportsValidate
 
 class FieldValidateIntention : PsiElementBaseIntentionAction() {
     override fun startInWriteAction(): Boolean = true
@@ -33,10 +31,7 @@ class FieldValidateIntention : PsiElementBaseIntentionAction() {
         val field = decoratedField.field
         val attributes = decoratedField.attributeListList.flatMap(FsdAttributeList::getAttributeList)
 
-        val isEnumTyped = field.type.referenceType?.reference?.resolve()?.parent is FsdEnumSpec
-        val isApplicableType = getValidateTemplateForField(field) != null || isEnumTyped
-
-        return isApplicableType && attributes.none { it.attributename.text == "validate" }
+        return supportsValidate(field) && attributes.none { it.attributename.text == "validate" }
     }
 
     override fun invoke(project: Project, editor: Editor?, element: PsiElement) {
@@ -87,25 +82,6 @@ class FieldValidateIntention : PsiElementBaseIntentionAction() {
             templateManager.startTemplate(editor, template)
         }
     }
-
-    private fun getValidateTemplateForField(field: FsdField): Template? {
-        val type = field.type.text
-
-        val templateKey = when {
-            type == "string" -> "svalid"
-            isCollection(type) -> "cvalid"
-            isNumber(type) -> "nvalid"
-            else -> return null
-        }
-
-        return TemplateSettings.getInstance().getTemplate(templateKey, FsdLanguage.displayName)
-    }
-
-    private fun isCollection(typeName: String) = typeName == "bytes" ||
-        typeName.startsWith("map<") ||
-        typeName.endsWith("[]")
-
-    private fun isNumber(typeName: String) = typeName == "int32" || typeName == "int64"
 
     companion object {
         val TEXT = FsdBundle.getMessage("intentions.validate.field.text")